refactor(roles): purge legacy 'researcher' and 'admin' role literals
Renames every role-string literal to 'member' or 'owner', deletes the stale
local ProjectWithRole union in types/backend.ts (single source of truth is
now the OpenAPI-generated type), renames isAdmin to isOwner in
ProjectMembersPage and ProjectSettingsPage, and renames the 'researcher'
test variable in test_rbac.py for semantic consistency.
