+++ frontend/src/pages/admin/ConcourseDetailPage.tsx
@@ -329,7 +329,7 @@ export default function ConcourseDetailPage() {
                                 parentId={concourse.id}
                                 currentUserId={currentUser.id}
                                 isOwner={projectRole === 'owner'}
-                                canEdit={projectRole === 'owner' || projectRole === 'researcher'}
+                                canEdit={projectRole === 'owner' || projectRole === 'member'}
                                 members={projectMembers}
                             />
                         )}
